Support for defaults methods in xml configuration
